@@ -13,18 +13,31 @@
     "bootstrap": "^4.6.0",
     "expo": "~40.0.0",
     "expo-status-bar": "~1.0.3",
-    "react": "16.13.1",
+    "prop-types": "^15.7.2",
+    "react": "^17.0.1",
     "react-bootstrap": "^1.5.0",
-    "react-dom": "16.13.1",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-40.0.1.tar.gz",
+    "react-dom": "^17.0.1",
+    "react-native": "^0.63.2",
+    "react-native-animated-component": "^0.1.0",
+    "react-native-animation-library": "0.0.8",
     "react-native-bootstrap-styles": "^4.5.0-r",
+    "react-native-button": "^3.0.1",
+    "react-native-cardview": "^2.0.5",
     "react-native-elements": "^3.2.0",
     "react-native-navbar": "^2.1.0",
+    "react-native-simple-animations": "^0.2.3",
     "react-native-snap-carousel": "^3.9.1",
-    "react-native-web": "^0.15.0"
+    "react-native-vector-icons": "^8.0.0",
+    "react-native-web": "~0.13.12"
   },
   "devDependencies": {
-    "@babel/core": "~7.9.0"
+    "@babel/core": "~7.9.0",
+    "@babel/preset-react": "^7.12.13",
+    "@types/expo": "^33.0.1",
+    "@types/react": "^17.0.0",
+    "@types/react-dom": "^17.0.0",
+    "@types/react-native": "^0.63.45",
+    "typescript": "^4.1.3"
   },
   "private": true
 }
